+++ packages/core/modules/Plugin.js
@@ -12,17 +12,19 @@ Kepler.Plugin = function(plugin) {
 	{
 		if(!this.plugins[plugin.name]) {
 
-			if(_.isObject(plugin.templates)) {
+			var ptmpls = plugin.templates;
 
-				for(var placeholder in plugin.templates) {
-					if(plugin.templates[placeholder] && !K.templates[placeholder])
+			if(_.isObject(ptmpls)) {
+
+				for(var placeholder in ptmpls) {
+					if(ptmpls[placeholder] && !_.has(K.templates,placeholder) )
 						K.templates[placeholder]= {};
 				}
-		
+
 				for(var placeholder in K.templates) {
-					if(plugin.templates[placeholder]) {
+					if(_.isObject(K.templates[placeholder]) && ptmpls[placeholder]) {
 						
-						var tt = K.Plugin.normalizePlacehoders(plugin.templates[placeholder], plugin.name);
+						var tt = K.Plugin.normalizePlaceholders(ptmpls[placeholder], plugin.name);
 						
 						_.extend(K.templates[placeholder], tt);
 					}
@@ -35,9 +37,8 @@ Kepler.Plugin = function(plugin) {
 			if(_.isObject(plugin.schemas))
 				_.deepExtend(K.schemas, plugin.schemas);
 
-			if(_.isObject(plugin.settings)){
+			if(_.isObject(plugin.settings))
 				_.deepExtend(K.settings, plugin.settings);
-			}
 
 			this.plugins[plugin.name] = plugin;
 		}
@@ -52,35 +53,36 @@ Kepler.Plugin = function(plugin) {
  * @param  {[type]} tt [description]
  * @return {[type]}    [description]
  */
-Kepler.Plugin.normalizePlacehoders = function(tt, plugin) {
+Kepler.Plugin.normalizePlaceholders = function(tmpls, plugin) {
 	
 	//TODO add new field 'replace'
 	
-	if(_.isString(tt))
-		tt = _.object([tt],[true]);
+	if(_.isString(tmpls))
+		tmpls = _.object([tmpls],[true]);
 	
-	for(var t in tt) {
-		if(tt[t]===true)
-			tt[t]= {};
+	for(var t in tmpls) {
+		if(tmpls[t]===true)
+			tmpls[t]= {};
 
-			if(_.isObject(tt[t])) {
-			tt[t]= _.defaults(tt[t], {
+			if(_.isObject(tmpls[t])) {
+			tmpls[t]= _.defaults(tmpls[t], {
 				plugin: plugin,
 				show: true,
 				order: 0
 			});
 		}
-		else if(tt[t]===false)
-			tt[t]= {show: false};
+		else if(tmpls[t]===false)
+			tmpls[t]= {show: false};
 	}
-	return tt;
+	return tmpls;
 };
 
 Kepler.Plugin.templatesByPlaceholder = function(placeholder, data, sep) {
 	
 	sep = sep || "\n";
 
-	var tmpls = [], sorts = [];
+	var tmpls = [],
+		sorts = [];
 
 	if(!placeholder 
 		//TODO uncomment and decide if check also it.. || !Template[placeholder]
@@ -88,8 +90,8 @@ Kepler.Plugin.templatesByPlaceholder = function(placeholder, data, sep) {
 
 	for(var parentName in K.templates) {
 
-
 		if(parentName === placeholder) {
+
 			//convert in array of objects
 			//TODO https://stackoverflow.com/questions/49611332/underscorejs-object-to-collection
 			var arr = _.map(K.templates[parentName], function(v,k) {
@@ -112,7 +114,6 @@ Kepler.Plugin.templatesByPlaceholder = function(placeholder, data, sep) {
 		}
 	}
 
-	
 	return tmpls;	
 };
 
@@ -122,21 +123,24 @@ if(Meteor.isClient) {
 
 		for(var placeholder in K.templates) {
 			
-			var tt = K.templates[placeholder];
+			var tmpls = K.templates[placeholder];
 
-			for(var t in tt) {
+			for(var t in tmpls) {
 				if(!Template[t]) {
-					console.warn('Template not exists: "'+t+'", defined in plugin: '+tt[t].plugin)
-					delete tt[t];
+					console.warn('Template not exists: "'+t+'", defined in plugin: '+tmpls[t].plugin)
+					delete tmpls[t];
 				}
 			}	
 		}
 
 		for(var placeholder in sets) {
 
+			if(sets[placeholder].show===false)
+				delete K.templates[placeholder];
+
 			if(!K.templates[placeholder]) continue;
 
-			_.extend(K.templates[placeholder], K.Plugin.normalizePlacehoders(sets[placeholder]) );
+			_.extend(K.templates[placeholder], K.Plugin.normalizePlaceholders(sets[placeholder]) );
 		}
 
 	});