+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 Complete list of releases [here](https://github.com/Keplerjs/Kepler/releases)
 
 ## Next version
+- disable placeholders by settings.json
 - fix panelUser template
 - block email send if user created by admin
 - fix profile messages sort